not allowing admin to delete reviews to ensure integrity of reviews

+++ src/main/java/bham/team/web/rest/ReviewResource.java
@@ -15,6 +15,9 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import tech.jhipster.web.util.HeaderUtil;
 import tech.jhipster.web.util.ResponseUtil;
@@ -162,6 +165,16 @@ public class ReviewResource {
      */
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteReview(@PathVariable("id") Long id) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        boolean isAdmin = false;
+        for (GrantedAuthority auth : authentication.getAuthorities()) {
+            if (auth.getAuthority().equals("ROLE_ADMIN")) {
+                isAdmin = true;
+            }
+        }
+        if (isAdmin) {
+            throw new BadRequestAlertException(" Only admins can create " + ENTITY_NAME, ENTITY_NAME, " idinvalid ");
+        }
         LOG.debug("REST request to delete Review : {}", id);
         reviewService.delete(id);
         return ResponseEntity.noContent()
